Finalize functional prototype with Wahoo Dongle
Resolves a threading issue where a delegate method
was being called on a background thread. This
caused an assertion by the OS as I was sending
the notification to a UI element on a background
thread.

This is more or less a testable state. I was able
to take this on a ride and it worked more or less
as advertized.