Add BTLE support
Set up the Bluetooth delegate and any
of the basic infrastructure to get us
to the point where we're seeing actual
power data. Currently tested from the
LightBlue app in the app store. It is
unclear if this is the way the data
would be sent over as there seems to
be a lot of noise. Testing with actual
hardware will close that loop.