Move every zone down by 1 slot
After connecting PowerIntervals to an actual
power meter, it became clear that the threshold
of the zones were off by 1. Active Recovery's
lower end is 0 watts. The first threshold we
encounter is the Endurance threshold. It makes
sense, then, to get rid of Active Recovery
as a zone. This caused a bunch of off by 1
adjustments which actually cleaned some things
up.