Fix bug where old power zones show up in set zone view
A bug arose when we attempted to show dummy data in
the chart for one of the blurry views. We really
only want to set dummy zone data when we haven't
already set zones for this user. We fix this by
simply checking if the zones have been set, if
they have not, then we create the dummy ones.

fixes #13